Make iOS trailing newlines not use current typing attributes (#183)
Trailing newlines on iOS need to be measured as if there was a fake `I`
character to get the proper measurement. But incorrectly, current typing
attributes were used for the character. This resulted in a funny thing;
trailing newline would change sizes if we changed selection from a, say,
heading line to non heading line. Now it uses the default typing
attributes.