fix: state change emitting with default values improvement (#251)
Fixes:
https://github.com/software-mansion/react-native-enriched/issues/237
Turns out `eventEmitter` isn't available from the very beginning, but
the iOS code was remembering recently active styles even if no state
change event was emitted. This way, when the event emitter was finally
available, no state change event was emitted because recently active
styles were exactly same (even though no event went out).
Except that, I added all lifecycle callbacks
(`anyTextMayHaveBeenModified` that e.g. runs the state change event) to
the `didMoveToWindow` because it seems like the event emitter is defined
then.
The issue was;`defaultValue` was set but its state changes emitted only
when some change was done to an input (and not on focus or when
eventEmitter is available). The PR fixes both scenarios.

---------

Co-authored-by: Igor Furgała <74370735+exploIF@users.noreply.github.com>