feat: unify initial text/html iOS emitting with Android (#270)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect -->

# Summary

After the latest PRs it also came to my mind that Android and iOS emit
the initial events a bit differently. If we have some `defaultValue`
set, iOS wouldn't emit text (and would emit html always, even if
`defaultValue` was not set), while Android properly emitted both html
and text only if the prop is set.

This PR changes: 
- initial `recentlyEmittedHtml` value so that it doesn't emit with no
`defaultValue`
- sets `recentlyEmittedString` only if the `onChangeText` event properly
wen through (event emitter was defined). This way text emits if it's
different than the initial `""`

## Test Plan

Run the example iOS app once with no `defaultValue` and the other time
with `defaultValue`. Observe that in the first scenario no initial
events are dispatched, while in the second they are emitted as soon as
emitter is ready.


## Screenshots / Videos

--

## Compatibility

| OS      | Implemented |
| ------- | :---------: |
| iOS     |    ✅     |
| Android |    ✅ (was fine already)    |