fix: iOS startMention check range (#278)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect -->

# Summary

The `startMention` ref method on iOS takes wrong range into
consideration when checking for blocks and conflicts. Can't be
reproduced in the current state of main branch, but on the codeblocks
feature branch I was not able to add an indicator this way because a
codeblock was detected within default `NSRange` value `(0,0)`.

This PR puts a proper `textView.selectedRange` range in there.

## Test Plan

--

## Screenshots / Videos

--

## Compatibility

| OS      | Implemented |
| ------- | :---------: |
| iOS     |    ✅     |
| Android |    (doesn't apply)    |