feat: proper iOS first renders measurements (#277)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect -->

# Summary

Fixes
https://github.com/software-mansion/react-native-enriched/issues/238 on
iOS.

First render on iOS was looking pretty choppy because there was no way
to measure the component on the very first render, because it's still
not defined at that time.
This PR adds a mock text input that gets created as soon as the Shadow
Node is created, gets the initial component's props and is used for
measurements before the real input is defined. After that the mock gets
unmounted (when the real input is finally defined).

## Before vs After

Please play the videos at x0.25 speed to notice the choppiness:

### Before:


https://github.com/user-attachments/assets/69aa524a-5ec1-4fbc-8220-610d2280eaf4

### After:


https://github.com/user-attachments/assets/8f07e5a1-67e5-4c6d-90df-d0b4f44e0f62

## Test Plan

Add a state that toggles input's visibility, with the default visibility
at `false`. Then toggle it and see properly measured component appear
with no intermediate states.

## Compatibility

| OS      | Implemented |
| ------- | :---------: |
| iOS     |    ✅    |
| Android |    ✅(already done)    |