fix: removing all characters from a line logic on iOS (#296)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect -->

# Summary

Fixes:
https://github.com/software-mansion/react-native-enriched/issues/287.
One of previous PRs has introduced a regression in the logic behind
removing all the characters from a line.
When it happens, iOS applies attributes from the previous line.
It's especially bad when paragraph styles are in the line before and a
bullet point suddenly appears.

The original fix was to manually remove the characters, remove typing
attributes and for the paragraph styles reapply the style that was there
in the current line.
Some PR must have removed the typing attributes/style reapplying part
when working with that code so I'm returning it to the properly working
state.

## Test Plan

Scenario 1:
Write one line with some paragraph style.
Then another line with no styling and remove all the characters from the
second line.
The paragraph style from the first line shouldn't expand to the second
line.

Scenario 2:
Write one line with some paragraph style.
Then another line with a different paragraph style and remove all the
characters from the second line.
The style should still be there but with an empty line and the style
from first paragraph shouldn't take over the second line's styling.

## Screenshots / Videos

### Before


https://github.com/user-attachments/assets/106c0ae8-f384-46a2-b7d7-3e7d833923e1

### After


https://github.com/user-attachments/assets/cfb8130d-c981-4d84-b015-518be3e6119f

## Compatibility

| OS      | Implemented |
| ------- | :---------: |
| iOS     |    ✅     |
| Android |    ❌     |