feat: do not parse spannable if it's not needed - Android (#307)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect -->

# Summary

Android only, already implemented on iOS. This PR ensures that spannable
is not parsed into HTML, when `onChangeHtml` is not provided, which
slightly improvements performance.

## Test Plan

Provide **clear steps so another contributor can reproduce the behavior
or verify the feature works**.
For example:

- add a native android log inside `EnrichedSpanWatcher.emitEvent`
- ensure `onChangeHtml` prop is provided
- notice that any text/style change triggers spannable -> HTML parsing
- remove `onChangeHtml`
- notice that spannable -> HTML is not parsed, as it's not needed

## Screenshots / Videos

n/a

## Compatibility

| OS      | Implemented |
| ------- | :---------: |
| iOS     |    ❌     |
| Android |    ✅     |