fix: toggling inline styles - Android (#313)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect -->

# Summary

- Fixes issue introduced in
https://github.com/software-mansion/react-native-enriched/pull/281

## Test Plan

Provide **clear steps so another contributor can reproduce the behavior
or verify the feature works**.
For example:

- Make one word bold.
- Make another, separate word bold.
- Create a selection starting from the middle of the first bold word to
the middle of the second bold word (the selection will include bold →
regular → bold segments).
- Call toggleBold.
- Notice that styles are properly applied

Also:
- set one word bold
- mark bold word
- toggle bold style and notice that style is properly removed

## Screenshots / Videos


https://github.com/user-attachments/assets/d98687c6-9b8d-4d9a-b157-9b6b07753dfe

## Compatibility

| OS      | Implemented |
| ------- | :---------: |
| iOS     |    ❌     |
| Android |    ✅     |