feat: get HTML value imperatively (#310)
# Summary

Adds the ability to use a getHTML method to get the current state of the
editor

This is incredibly important to my team since we're adding heavy usage
of this library into our app and launching the update in about two
weeks. The current performance with onChangeHTML is inadequate for us,
so this change is very important.

## Test Plan

Simplest options is to create an enriched input with a ref, and a button
below it that triggers a function which calls getHTML() and logs it out
to the console. You should see the logged HTML.
## Compatibility

| OS      | Implemented |
| ------- | :---------: |
| iOS     |    ✅     |
| Android |    ✅     |

---------

Co-authored-by: Ahmed Haracic <ahmed@Ahmeds-MacBook-Pro.local>
Co-authored-by: Mikołaj Szydłowski <9szydlowski9@gmail.com>
Co-authored-by: Igor Furgala <exploif@icloud.com>