fix: recalculating layout after changing html styles - Android (#233)
### Description

If we change the HTML styles on Android, they will not be applied in any
way until we first enter any styles. We also got a bug with the initial
defaultValue, because everything started to depend on who was the first
to be counted (it depends on who was the first to be passed from JS to
the component).
Thus, I added the calculation of the lead when changing the HTML style
on Android.


### Videos

Before:


https://github.com/user-attachments/assets/0ccaae0c-d5e0-4270-b000-534260803ea3



https://github.com/user-attachments/assets/e89536be-c787-477a-8aba-fba2e0a6919d



After:


https://github.com/user-attachments/assets/6d27d9f7-cbbf-40f3-9010-557298a3eaf7

---------

Co-authored-by: Igor Furgala <exploif@icloud.com>
Co-authored-by: Igor Furgała <74370735+exploIF@users.noreply.github.com>