fix(iOS): draw elements only in visible area (#325)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect -->

# Summary

Currently, we redraw elements in the whole input range. There is no
sense in redrawing elements that are hidden(not visible to the user),
and it impacts performance
1. Initial mount
2. Scroll performance


## Test Plan

1. Test scroll/mount on iOS

## Screenshots / Videos

### Before


https://github.com/user-attachments/assets/43f6d113-f676-47a7-b863-b7797df4e9ca

### After


https://github.com/user-attachments/assets/e87334f1-e074-4bee-98f6-0d59768fa557

Include any visual proof that helps reviewers understand the change — UI
updates, bug reproduction or the result of the fix.

## Compatibility

| OS      | Implemented |
| ------- | :---------: |
| iOS     |    ✅    |
| Android |    ❌     |

---------

Co-authored-by: szydlovsky <9szydlowski9@gmail.com>