fix(iOS): do not set the selection on the link when applying the initial styles (#314)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect -->

# Summary

Added `withSelection` argument for `addLink` method to not set selection
when we are applying initial input.

Basically, it solves 2 issues

1. When we initially apply styles, it will always set the selection to
the latest link in HTML
2. If we replace the HTML value, it will execute the code in the
following order

`applyProcessedStyles -> addLink -> setSelection ->
textViewDidChangeSelection -> tryUpdatingActiveStyles`

It is not a big problem, but if we process a really huge HTML that
contains a lot of links, it will be a bottleneck

## Test Plan

1. Set the default value like: `<html><h1>Link test</h1><p><a
href="test.com">test link</a></p></html>`
2. Open the app
3. Observe that the focus is set correctly

## Screenshots / Videos

### Before:


https://github.com/user-attachments/assets/38d9593a-7a6d-4dcb-976e-0834893384fa

### After:


https://github.com/user-attachments/assets/7126719e-0e73-42de-b42e-d8788f04ff52



Include any visual proof that helps reviewers understand the change — UI
updates, bug reproduction or the result of the fix.

## Compatibility

| OS      | Implemented |
| ------- | :---------: |
| iOS     |    ✅     |
| Android |    ❌     |

Co-authored-by: Mikołaj Szydłowski <9szydlowski9@gmail.com>